reuse platforms filter on /platforms
