+++ docs/platforms/index.mdx
@@ -20,9 +20,7 @@ We also <Link to="https://sentry.zendesk.com/hc/en-us/articles/23961784367387">d
 
 </Note>
 
-## Platforms
-
-<PlatformGrid />
+<PlatformFilter />
 
 ## Community Supported
