+++ src/mdxComponents.ts
@@ -20,6 +20,7 @@ import {PageGrid} from './components/pageGrid';
 import {ParamTable} from './components/paramTable';
 import {PiiFields} from './components/piiFields';
 import {PlatformCategorySection} from './components/platformCategorySection';
+import {PlatformFilter} from './components/platformFilter';
 import {PlatformGrid} from './components/platformGrid';
 import {PlatformIdentifier} from './components/platformIdentifier';
 import {PlatformLink} from './components/platformLink';
@@ -42,6 +43,7 @@ export function mdxComponents(
     Arcade,
     Break,
     CliChecksumTable,
+    PlatformFilter,
     CodeBlock,
     CodeTabs,
     ConfigKey,
